// src/backend-admin-sdk/src/models/PickupAddressAdminInput.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OPENCOURIER-API
 * OpenCourier Backend API
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';

export const EnumCountryCodeAdmin = {
    Us: 'US',
    Gr: 'GR',
    Ca: 'CA',
    Gb: 'GB',
    Au: 'AU',
    Mx: 'MX',
} as const;
export type EnumCountryCodeAdmin = typeof EnumCountryCodeAdmin[keyof typeof EnumCountryCodeAdmin];

/**
 * 
 * @export
 * @interface PickupAddressAdminInput
 */
export interface PickupAddressAdminInput {
    /**
     * 
     * @type {string}
     * @memberof PickupAddressAdminInput
     */
    street: string;
    /**
     * 
     * @type {string}
     * @memberof PickupAddressAdminInput
     */
    houseNumber?: string;
    /**
     * 
     * @type {string}
     * @memberof PickupAddressAdminInput
     */
    city: string;
    /**
     * 
     * @type {string}
     * @memberof PickupAddressAdminInput
     */
    state?: string;
    /**
     * 
     * @type {string}
     * @memberof PickupAddressAdminInput
     */
    zipCode?: string;
    /**
     * 
     * @type {EnumCountryCodeAdmin}
     * @memberof PickupAddressAdminInput
     */
    countryCode: EnumCountryCodeAdmin;
    /**
     * 
     * @type {number}
     * @memberof PickupAddressAdminInput
     */
    latitude: number;
    /**
     * 
     * @type {number}
     * @memberof PickupAddressAdminInput
     */
    longitude: number;
    /**
     * 
     * @type {string}
     * @memberof PickupAddressAdminInput
     */
    formattedAddress?: string;
}

/**
 * Check if a given object implements the PickupAddressAdminInput interface.
 */
export function instanceOfPickupAddressAdminInput(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "street" in value;
    isInstance = isInstance && "city" in value;
    isInstance = isInstance && "countryCode" in value;
    isInstance = isInstance && "latitude" in value;
    isInstance = isInstance && "longitude" in value;

    return isInstance;
}

export function PickupAddressAdminInputFromJSON(json: any): PickupAddressAdminInput {
    return PickupAddressAdminInputFromJSONTyped(json, false);
}

export function PickupAddressAdminInputFromJSONTyped(json: any, ignoreDiscriminator: boolean): PickupAddressAdminInput {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        'street': json['street'],
        'houseNumber': !exists(json, 'houseNumber') ? undefined : json['houseNumber'],
        'city': json['city'],
        'state': !exists(json, 'state') ? undefined : json['state'],
        'zipCode': !exists(json, 'zipCode') ? undefined : json['zipCode'],
        'countryCode': json['countryCode'],
        'latitude': json['latitude'],
        'longitude': json['longitude'],
        'formattedAddress': !exists(json, 'formattedAddress') ? undefined : json['formattedAddress'],
    };
}

export function PickupAddressAdminInputToJSON(value?: PickupAddressAdminInput | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        'street': value.street,
        'houseNumber': value.houseNumber,
        'city': value.city,
        'state': value.state,
        'zipCode': value.zipCode,
        'countryCode': value.countryCode,
        'latitude': value.latitude,
        'longitude': value.longitude,
        'formattedAddress': value.formattedAddress,
    };
}