// src/backend-admin-sdk/src/models/PartnerCreateAdminInput.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OPENCOURIER-API
 * OpenCourier Backend API
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { PickupAddressAdminInput } from './PickupAddressAdminInput';
import {
    PickupAddressAdminInputFromJSON,
    PickupAddressAdminInputFromJSONTyped,
    PickupAddressAdminInputToJSON,
} from './PickupAddressAdminInput';

/**
 * 
 * @export
 * @interface PartnerCreateAdminInput
 */
export interface PartnerCreateAdminInput {
    /**
     * 
     * @type {string}
     * @memberof PartnerCreateAdminInput
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof PartnerCreateAdminInput
     */
    phoneNumber?: string;
    /**
     * 
     * @type {string}
     * @memberof PartnerCreateAdminInput
     */
    email: string;
    /**
     * 
     * @type {string}
     * @memberof PartnerCreateAdminInput
     */
    password: string;
    /**
     * 
     * @type {PickupAddressAdminInput}
     * @memberof PartnerCreateAdminInput
     */
    pickupAddress?: PickupAddressAdminInput;
}

/**
 * Check if a given object implements the PartnerCreateAdminInput interface.
 */
export function instanceOfPartnerCreateAdminInput(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "name" in value;
    isInstance = isInstance && "email" in value;
    isInstance = isInstance && "password" in value;

    return isInstance;
}

export function PartnerCreateAdminInputFromJSON(json: any): PartnerCreateAdminInput {
    return PartnerCreateAdminInputFromJSONTyped(json, false);
}

export function PartnerCreateAdminInputFromJSONTyped(json: any, ignoreDiscriminator: boolean): PartnerCreateAdminInput {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        'name': json['name'],
        'phoneNumber': !exists(json, 'phoneNumber') ? undefined : json['phoneNumber'],
        'email': json['email'],
        'password': json['password'],
        'pickupAddress': !exists(json, 'pickupAddress') ? undefined : PickupAddressAdminInputFromJSON(json['pickupAddress']),
    };
}

export function PartnerCreateAdminInputToJSON(value?: PartnerCreateAdminInput | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        'name': value.name,
        'phoneNumber': value.phoneNumber,
        'email': value.email,
        'password': value.password,
        'pickupAddress': PickupAddressAdminInputToJSON(value.pickupAddress),
    };
}