// src/backend-admin-sdk/src/models/PartnerUpdateAdminInput.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OPENCOURIER-API
 * OpenCourier Backend API
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { PickupAddressAdminInput } from './PickupAddressAdminInput';
import {
    PickupAddressAdminInputFromJSON,
    PickupAddressAdminInputFromJSONTyped,
    PickupAddressAdminInputToJSON,
} from './PickupAddressAdminInput';

/**
 * 
 * @export
 * @interface PartnerUpdateAdminInput
 */
export interface PartnerUpdateAdminInput {
    /**
     * 
     * @type {string}
     * @memberof PartnerUpdateAdminInput
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof PartnerUpdateAdminInput
     */
    phoneNumber?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PartnerUpdateAdminInput
     */
    logo?: string | null;
    /**
     * 
     * @type {string}
     * @memberof PartnerUpdateAdminInput
     */
    webhookUrl?: string | null;
    /**
     * 
     * @type {PickupAddressAdminInput}
     * @memberof PartnerUpdateAdminInput
     */
    pickupAddress?: PickupAddressAdminInput;
}

/**
 * Check if a given object implements the PartnerUpdateAdminInput interface.
 */
export function instanceOfPartnerUpdateAdminInput(value: object): boolean {
    return true;
}

export function PartnerUpdateAdminInputFromJSON(json: any): PartnerUpdateAdminInput {
    return PartnerUpdateAdminInputFromJSONTyped(json, false);
}

export function PartnerUpdateAdminInputFromJSONTyped(json: any, ignoreDiscriminator: boolean): PartnerUpdateAdminInput {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        'name': !exists(json, 'name') ? undefined : json['name'],
        'phoneNumber': !exists(json, 'phoneNumber') ? undefined : json['phoneNumber'],
        'logo': !exists(json, 'logo') ? undefined : json['logo'],
        'webhookUrl': !exists(json, 'webhookUrl') ? undefined : json['webhookUrl'],
        'pickupAddress': !exists(json, 'pickupAddress') ? undefined : PickupAddressAdminInputFromJSON(json['pickupAddress']),
    };
}

export function PartnerUpdateAdminInputToJSON(value?: PartnerUpdateAdminInput | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        'name': value.name,
        'phoneNumber': value.phoneNumber,
        'logo': value.logo,
        'webhookUrl': value.webhookUrl,
        'pickupAddress': PickupAddressAdminInputToJSON(value.pickupAddress),
    };
}